Update Conjure parameters
Updates Conjure parameters to embed name as part of the parameter.
Also changes the definition of Conjure config structs to be a slice
of named project configs rather than a map that is ordered alphabetically
after the fact. However, the wire format remains the same (regular map).

The primary functional difference is that, in any instance in which Conjure
parameters are processed, the ordering is now based on the parameter ordering
(which, in standard usage, mirrors the map key order in the source YAML
configuration) rather than alphabetically by project name.