Update Conjure parameters (#679)
Updates Conjure parameters to embed name as part of the parameter. Also changes the definition of Conjure config structs to be a slice of named project configs rather than a map that is ordered alphabetically after the fact. However, the wire format remains the same (regular map).

Also updates verification errors to print the name of the Conjure project for which verification failed rather than the integer index of the project.